be/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
mysqlclient>=2.1
django-cors-headers>=3.14
gunicorn>=20.1
djangorestframework-simplejwt>=5.2
whitenoise